cmd: omit namespace in service discovery
This query for the `*` namespace only returns results from
namespaces that the Nomad token has explicit access to.

By default, the workload Nomad token provided by the `identity`
block has read access to all service discovery information,
but does not have explicit access to any namespaces,
so this query returns nothing when using it.

Therefore, omit querying for the `*` namespace.